+++ src/main/java/subway/persistence/SectionRepository.java
@@ -1,4 +1,4 @@
-package subway.repository;
+package subway.persistence;
 
 import subway.domain.Section;
