+++ src/main/java/subway/persistence/dao/SectionDao.java
@@ -1,4 +1,4 @@
-package subway.repository.dao;
+package subway.persistence.dao;
 
 import java.util.HashMap;
 import java.util.Map;
