+++ src/main/java/subway/persistence/dao/StationDao.java
@@ -1,4 +1,4 @@
-package subway.repository.dao;
+package subway.persistence.dao;
 
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
