+++ src/test/java/subway/persistence/LineRepositoryTest.java
@@ -1,4 +1,4 @@
-package subway.repository;
+package subway.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
